requirements.txt
aiofile
aiofiles
fastapi
httpx
pyyaml
uvicorn